Fix the OpenSextant image.
There were a couple problems. The most embarrassing one is that the Dockerfile
was building from a tagged release in Github, so it wasn't using any of
the updates in the repo. I tagged a new release.

Also, we need to convince `ant` to use Java 1.8 by patching `build-classpath`
and creating a few directories neglected by the 1.8.0-openjdk packages.